+++ backend_users/tomato/api/auth.py
@@ -1,0 +1,5 @@
+from _shared import _getUser
+
+def user_check_password(name, password):
+	user = _getUser(name)
+	return user.checkPassword(password)
