+++ backend_users/tomato/api/organization.py
@@ -1,0 +1,23 @@
+from ..organization import Organization
+from _shared import _getOrganization
+
+def organization_create(**args):
+	org = Organization.create(**args)
+	return org.name
+
+def organization_list():
+	return [o.info() for o in Organization.objects.all()]
+
+def organization_info(name):
+	orga = _getOrganization(name)
+	return orga.info()
+
+def organization_modify(name, **args):
+	orga = _getOrganization(name)
+	orga.modify(**args)
+	return orga.info()
+
+def organization_remove(name):
+	orga = _getOrganization(name)
+	orga.remove()
+	return True
